use module feature injection for matrixReducer and matrixEffects and therefore remove app.reducer and app.effects
