+++ src/app/app.state.ts
@@ -1,0 +1,5 @@
+import * as fromMatrix from './matrix/matrix.reducers';
+
+export interface AppState {
+  [fromMatrix.matrixFeatureKey]: fromMatrix.MatrixState;
+}
