--- src/app/store/app.effects.ts
@@ -1,3 +1,0 @@
-import { MatrixEffects } from '../matrix/matrix.effects';
-
-export const appEffects = [MatrixEffects];
