--- src/app/store/app.state.ts
@@ -1,5 +1,0 @@
-import { MatrixState } from '../matrix/matrix.interfaces';
-
-export interface AppState {
-  matrix: MatrixState;
-}
